fix: useSession to show user information on user nav
