+++ package.json
@@ -60,6 +60,6 @@
     "@types/node": "18.11.18",
     "@types/react": "18.0.27",
     "@types/react-dom": "18.0.10",
-    "typescript": "5.0.4"
+    "typescript": "5.1.3"
   }
 }
