--- src/types/modul-page.ts
@@ -1,8 +1,0 @@
-export interface IModulParams {
-  params: {
-    modulId: string | number;
-  };
-  searchParams?: {
-    presensiId: string | number;
-  };
-}
