--- src/types/page/materi.ts
@@ -1,8 +1,0 @@
-export interface IMateriParams {
-  params: {
-    modulId: string | number;
-  };
-  searchParams?: {
-    materiId: string | number;
-  };
-}
